Move integration tests to afkak.test.int
This allows them to easily be targeted with Trial, like:

    trial afkak.test.int

Otherwise you'd need to do something like (untested):

    trial afkak/test/test_*_integration.py

The "int" module name is unfortunate, but it aligns with the suite
naming used in tox.ini.
