+++ afkak/test/int/test_consumer_integration.py
@@ -20,15 +20,15 @@ from nose.twistedtools import deferred
 from twisted.internet.defer import inlineCallbacks, returnValue
 from twisted.trial import unittest
 
-from .. import Consumer, create_message
-from ..common import (
+from afkak import Consumer, create_message
+from afkak.common import (
     OFFSET_COMMITTED, OFFSET_EARLIEST, ConsumerFetchSizeTooSmall,
     ProduceRequest, RetriableBrokerResponseError,
 )
-from ..consumer import FETCH_BUFFER_SIZE_BYTES
-from .testutil import (
-    IntegrationMixin, async_delay, kafka_versions, random_string,
-)
+from afkak.consumer import FETCH_BUFFER_SIZE_BYTES
+from afkak.test.testutil import async_delay, random_string
+
+from .intutil import IntegrationMixin, kafka_versions
 
 log = logging.getLogger(__name__)
